Patch

+++ kernel/io/string.h
@@ -1,7 +1,7 @@
 #ifndef STRING_H
 #define STRING_H
 
-uint16_t strlen(char* c);
+size_t strlen(char* c);
 char digit_to_char(uint8_t input);
 char* substring(char* str, uint16_t start, uint16_t end);
 char* concat(char* str1, char* str2);
